Re-organize and begin adding comments
